integrate package-lint, provide a clean summary report
Fix #19
Fix #20

With this update all the validator functions have been
reimplemented to report lists of lines containing errors.
Those lists are consolidated per file and sorted, so that
the output now present line-by-line errors for each file.

Also removes most of the visually noisy dividers from the
output because they are not necessary any more. The only
markers in the output other than actual issues to be
addressed is now an "OK" or a "FAIL", i.e.

  filename.el OK
  filename.el FAIL

With this change the minimum requirement is now Emacs 24.4