retry logic for gpg key retrieval
The retry logic seems necessary to resolve intermittent failures
accessing keyservers.  We're naming Ubuntu's keyserver rather than
allowing gnupg to select from a pool, in hopes that the commercial
server is more consistently maintained than the members of the pool.
Three retries ought to be enough to receive the key, else fail.