handle `package-lint` error: package X not installable
Because this check uses the `package` library to validate when
dependencies can be installed from an available package repository,
we need to suppress the "package is not installable" errors when
running in a context where the `package` library is not initialized.

We're using the value of package-archive-contents to determine this.
When the value is nil, then we suppress "not installable" errors.